+++ src/pytest_workflow/util.py
@@ -1,9 +1,8 @@
+import hashlib
 import os
 import re
-import shutil
 import warnings
 from pathlib import Path
-from typing import List
 
 
 # This function was created to ensure the same conversion is used throughout
@@ -16,11 +15,6 @@ def replace_whitespace(string: str, replace_with: str = '_') -> str:
     :return: The string with whitespace converted.
     """
     return re.sub(r'\s+', replace_with, string)
-
-
-def rm_dirs(directories: List[Path]):
-    for directory in directories:
-        shutil.rmtree(str(directory))
 
 
 def is_in_dir(child: Path, parent: Path, strict: bool = False) -> bool:
@@ -59,7 +53,22 @@ def link_tree(src: Path, dest: Path) -> None:
         for path in os.listdir(str(src)):
             link_tree(Path(src, path), Path(dest, path))
     elif src.is_file() or src.is_symlink():
-        os.symlink(str(src), str(dest), target_is_directory=False)
+        dest.symlink_to(src, target_is_directory=False)
     else:  # Only copy files and symlinks, no devices etc.
         warnings.warn(f"Unsupported filetype. Skipping copying: '{str(src)}' "
                       f"to '{str(dest)}'.")
+
+
+# block_size 64k with python is a few percent faster than linux native md5sum.
+def file_md5sum(filepath: Path, block_size=64 * 1024) -> str:
+    """
+    Generates a md5sum for a file. Reads file in blocks to save memory.
+    :param filepath: a pathlib. Path to the file
+    :param block_size: Block size in bytes
+    :return: a md5sum as hexadecimal string.
+    """
+    hasher = hashlib.md5()  # nosec: only used for file integrity
+    with filepath.open('rb') as file_handler:  # Read the file in bytes
+        for block in iter(lambda: file_handler.read(block_size), b''):
+            hasher.update(block)
+    return hasher.hexdigest()