+++ tests/test_utils.py
@@ -13,13 +13,15 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with pytest-workflow.  If not, see <https://www.gnu.org/licenses/
+import hashlib
 import shutil
 import tempfile
 from pathlib import Path
 
 import pytest
 
-from pytest_workflow.util import is_in_dir, link_tree, replace_whitespace
+from pytest_workflow.util import file_md5sum, is_in_dir, link_tree, \
+    replace_whitespace
 
 WHITESPACE_TESTS = [
     ("bla\nbla", "bla_bla"),
@@ -49,7 +51,7 @@ def test_is_in_dir(child: str, parent: str, in_dir: bool):
 
 
 def test_link_tree():
-    pipelines_dir = Path(__file__).parent / Path("pipelines")
+    pipelines_dir = Path(__file__).parent / "pipelines"
     tempdir = Path(tempfile.mkdtemp(), "test")
     link_tree(pipelines_dir, tempdir)
     snakemake = Path(tempdir, "snakemake")
@@ -83,3 +85,14 @@ def test_link_tree_warning():
         link_tree(urandom, dest)
     # Destination should not be created
     assert not dest.exists()
+
+
+HASH_FILE_DIR = Path(__file__).parent / "hash_files"
+
+
+@pytest.mark.parametrize("hash_file", HASH_FILE_DIR.iterdir())
+def test_file_md5sum(hash_file: Path):
+    # No sec added because this hash is only used for checking file integrity
+    whole_file_md5 = hashlib.md5(hash_file.read_bytes()).hexdigest()  # nosec
+    per_line_md5 = file_md5sum(hash_file)
+    assert whole_file_md5 == per_line_md5